tic-tac

+++ optimization/src/main/java/com/gt/ml/pb/CountOnesTime.java
@@ -57,7 +57,7 @@ public class CountOnesTime {
 	        RandomizedHillClimbing rhc = new RandomizedHillClimbing(hcp);
 	        TimeTrainer fit = new TimeTrainer(rhc, time);
 	        fit.train();	
-	        int iterations = fit.getIterations();
+	        long iterations = fit.getIterations();
 	        Instance optimal = rhc.getOptimal();
 	        double optimalValue = ef.value(optimal);
 	        sumOptimalValue += optimalValue;
@@ -79,7 +79,7 @@ public class CountOnesTime {
 	        SimulatedAnnealing sa = new SimulatedAnnealing(100, .95, hcp);	        
 	        TimeTrainer fit = new TimeTrainer(sa, time);
 	        fit.train();	
-	        int iterations = fit.getIterations();
+	        long iterations = fit.getIterations();
 	        Instance optimal = sa.getOptimal();
 	        double optimalValue = ef.value(optimal);
 	        sumOptimalValue += optimalValue;
@@ -102,7 +102,7 @@ public class CountOnesTime {
 	        StandardGeneticAlgorithm ga = new StandardGeneticAlgorithm(200, 150, 25, gap);	        
 	        TimeTrainer fit = new TimeTrainer(ga, time);
 	        fit.train();	
-	        int iterations = fit.getIterations();
+	        long iterations = fit.getIterations();
 	        Instance optimal = ga.getOptimal();
 	        double optimalValue = ef.value(optimal);
 	        sumOptimalValue += optimalValue;
@@ -124,7 +124,7 @@ public class CountOnesTime {
 	        MIMIC mimic = new MIMIC(200, 100, pop);	        
 	        TimeTrainer fit = new TimeTrainer(mimic, time);
 	        fit.train();	
-	        int iterations = fit.getIterations();
+	        long iterations = fit.getIterations();
 	        Instance optimal = mimic.getOptimal();
 	        double optimalValue = ef.value(optimal);
 	        sumOptimalValue += optimalValue;
